# server/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "BISense API",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/standards": {
   "get": {
    "tags": [
     "standards"
    ],
    "summary": "List Standards",
    "operationId": "list_standards_api_standards_get",
    "parameters": [
     {
      "name": "q",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "maxLength": 200
        },
        {
         "type": "null"
        }
       ],
       "title": "Q"
      }
     },
     {
      "name": "kind",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "pattern": "^(standard|guidance|order|catalogue)(,(standard|guidance|order|catalogue))*$"
        },
        {
         "type": "null"
        }
       ],
       "title": "Kind"
      }
     },
     {
      "name": "category",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "maxLength": 200
        },
        {
         "type": "null"
        }
       ],
       "title": "Category"
      }
     },
     {
      "name": "compulsory",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "pattern": "^(yes|denotified|unknown|no)$"
        },
        {
         "type": "null"
        }
       ],
       "title": "Compulsory"
      }
     },
     {
      "name": "tier",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "pattern": "^[ABC]$"
        },
        {
         "type": "null"
        }
       ],
       "title": "Tier"
      }
     },
     {
      "name": "year_from",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer",
         "maximum": 2100,
         "minimum": 1900
        },
        {
         "type": "null"
        }
       ],
       "title": "Year From"
      }
     },
     {
      "name": "year_to",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer",
         "maximum": 2100,
         "minimum": 1900
        },
        {
         "type": "null"
        }
       ],
       "title": "Year To"
      }
     },
     {
      "name": "sort",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "pattern": "^(relevance|number|title|year)$",
       "default": "relevance",
       "title": "Sort"
      }
     },
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 1000,
       "minimum": 1,
       "default": 1,
       "title": "Page"
      }
     },
     {
      "name": "page_size",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 100,
       "minimum": 1,
       "default": 25,
       "title": "Page Size"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/StandardListOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/standards/suggest": {
   "get": {
    "tags": [
     "standards"
    ],
    "summary": "Suggest",
    "operationId": "suggest_api_standards_suggest_get",
    "parameters": [
     {
      "name": "q",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "minLength": 1,
       "maxLength": 100,
       "title": "Q"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/SuggestItem"
         },
         "title": "Response Suggest Api Standards Suggest Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/standards/{slug}": {
   "get": {
    "tags": [
     "standards"
    ],
    "summary": "Standard Detail",
    "operationId": "standard_detail_api_standards__slug__get",
    "parameters": [
     {
      "name": "slug",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Slug"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/StandardDetail"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/standards/{slug}/clauses": {
   "get": {
    "tags": [
     "standards"
    ],
    "summary": "All Clauses",
    "description": "Every clause of a document, in order, with verbatim text (the explorer's clause viewer).",
    "operationId": "all_clauses_api_standards__slug__clauses_get",
    "parameters": [
     {
      "name": "slug",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Slug"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/ClauseOut"
         },
         "title": "Response All Clauses Api Standards  Slug  Clauses Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/standards/{slug}/clauses/{number}": {
   "get": {
    "tags": [
     "standards"
    ],
    "summary": "Clause",
    "operationId": "clause_api_standards__slug__clauses__number__get",
    "parameters": [
     {
      "name": "slug",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Slug"
      }
     },
     {
      "name": "number",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Number"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ClauseOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/standards/{slug}/requirements": {
   "get": {
    "tags": [
     "standards"
    ],
    "summary": "Requirements",
    "operationId": "requirements_api_standards__slug__requirements_get",
    "parameters": [
     {
      "name": "slug",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Slug"
      }
     },
     {
      "name": "modality",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "pattern": "^(shall|shall_not|should|should_not|may|must)(,(shall|shall_not|should|should_not|may|must))*$"
        },
        {
         "type": "null"
        }
       ],
       "title": "Modality"
      }
     },
     {
      "name": "kind",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "maxLength": 200,
         "pattern": "^[a-z_,]+$"
        },
        {
         "type": "null"
        }
       ],
       "title": "Kind"
      }
     },
     {
      "name": "q",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "maxLength": 200
        },
        {
         "type": "null"
        }
       ],
       "title": "Q"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/RequirementsOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/standards/{slug}/requirements/plain": {
   "get": {
    "tags": [
     "standards"
    ],
    "summary": "Requirements Plain",
    "description": "AI interpretations of requirement statements (labelled as such in the UI).",
    "operationId": "requirements_plain_api_standards__slug__requirements_plain_get",
    "parameters": [
     {
      "name": "slug",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Slug"
      }
     },
     {
      "name": "lang",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "pattern": "^(en|hi|kn)$",
       "default": "en",
       "title": "Lang"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": {
          "type": "object",
          "additionalProperties": {
           "type": "string"
          }
         },
         "title": "Response Requirements Plain Api Standards  Slug  Requirements Plain Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "StandardListOut": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/StandardSummary"
      },
      "type": "array",
      "title": "Items"
     },
     "total": {
      "type": "integer",
      "title": "Total"
     },
     "page": {
      "type": "integer",
      "title": "Page"
     },
     "page_size": {
      "type": "integer",
      "title": "Page Size"
     },
     "facets": {
      "additionalProperties": {
       "items": {
        "$ref": "#/components/schemas/CategoryCount"
       },
       "type": "array"
      },
      "type": "object",
      "title": "Facets"
     }
    },
    "type": "object",
    "required": [
     "items",
     "total",
     "page",
     "page_size",
     "facets"
    ],
    "title": "StandardListOut"
   },
...
   "SuggestItem": {
    "properties": {
     "slug": {
      "type": "string",
      "title": "Slug"
     },
     "number": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Number"
     },
     "title": {
      "type": "string",
      "title": "Title"
     },
     "kind": {
      "type": "string",
      "enum": [
       "standard",
       "guidance",
       "order",
       "catalogue"
      ],
      "title": "Kind"
     }
    },
    "type": "object",
    "required": [
     "slug",
     "number",
     "title",
     "kind"
    ],
    "title": "SuggestItem"
   },
...
  }
 }
}